+++ trailblaze-agent/src/main/java/xyz/block/trailblaze/agent/util/LogHelper.kt
@@ -1,10 +1,10 @@
-package xyz.block.trailblaze.openai
+package xyz.block.trailblaze.agent.util
 
-import xyz.block.trailblaze.agent.model.TestObjective.TrailblazeObjective.TrailblazePrompt
+import xyz.block.trailblaze.agent.model.TestObjective
 import xyz.block.trailblaze.agent.model.TrailblazePromptStep
 
-object LogHelper {
-  fun logPromptStart(prompt: TrailblazePrompt) {
+internal object LogHelper {
+  fun logPromptStart(prompt: TestObjective.TrailblazeObjective.TrailblazePrompt) {
     println("\n\u001B[1;35m━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\u001B[0m")
     println("\u001B[1;35m🚀 Starting new agent objective: ${prompt.fullPrompt}\u001B[0m")
     println("\u001B[1;35m━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\u001B[0m")
