+++ trailblaze-common/dependencies/runtimeClasspath.txt
@@ -20,6 +20,7 @@ com.squareup.okio:okio-jvm:3.10.2
 com.squareup.okio:okio:3.10.2
 dev.mobile:maestro-client:1.40.0
 dev.mobile:maestro-orchestra-models:1.40.0
+dev.mobile:maestro-orchestra:1.40.0
 io.ktor:ktor-client-core-jvm:3.1.3
 io.ktor:ktor-client-core:3.1.3
 io.ktor:ktor-client-okhttp-jvm:3.1.3
